// SKYNET server/GC/730/generated/cs2.ts
export interface GcRoute<TRequest, TResponse> {
    readonly requestId: number;
    readonly request: ProtoDescriptor<TRequest>;
    readonly responseId: number;
    readonly response: ProtoDescriptor<TResponse>;
}

export interface EmptyRequest {
}

export type MessageId = number;
// ...
export interface CMsgGCCStrike15v2AccountRequestCoPlays {
    readonly players?: CMsgGCCStrike15v2AccountRequestCoPlays_Player[];
    readonly servertime?: number;
}

export interface CMsgGCCStrike15v2AcknowledgePenalty {
    readonly acknowledged?: number;
}

export interface CMsgGCCStrike15v2ClientGCRankUpdate {
    readonly rankings?: PlayerRankingInfo[];
}

export interface CMsgGCCStrike15v2ClientPlayerDecalSign {
    readonly data?: PlayerDecalDigitalSignature;
    readonly itemid?: bigint;
}

export interface CMsgGCCStrike15v2ClientRequestPlayersProfile {
    readonly requestIdDeprecated?: number;
    readonly accountIdsDeprecated?: number[];
    readonly accountId?: number;
    readonly requestLevel?: number;
}

export interface CMsgGCCStrike15v2GetEventFavoritesRequest {
    readonly allEvents?: boolean;
}

export interface CMsgGCCStrike15v2GetEventFavoritesResponse {
    readonly allEvents?: boolean;
    readonly jsonFavorites?: string;
    readonly jsonFeatured?: string;
}

export interface CMsgGCCStrike15v2MatchEndRunRewardDrops {
    readonly serverinfo?: CMsgGCCStrike15v2MatchmakingServerReservationResponse;
    readonly matchEndQuestData?: CMsgGCServerQuestUpdateData;
}

export interface CMsgGCCStrike15v2MatchList {
    readonly msgrequestid?: number;
    readonly accountid?: number;
    readonly servertime?: number;
    readonly matches?: CDataGCCStrike15v2MatchInfo[];
    readonly streams?: TournamentTeam[];
    readonly tournamentinfo?: CDataGCCStrike15v2TournamentInfo;
}

export interface CMsgGCCStrike15v2MatchListRequestRecentUserGames {
    readonly accountid?: number;
}

export interface CMsgGCCStrike15v2MatchListRequestTournamentGames {
    readonly eventid?: number;
}

export interface CMsgGCCStrike15v2MatchmakingClient2GCHello {
}

export interface CMsgGCCStrike15v2MatchmakingGC2ClientHello {
    readonly accountId?: number;
    readonly ongoingmatch?: CMsgGCCStrike15v2MatchmakingGC2ClientReserve;
    readonly globalStats?: GlobalStatistics;
    readonly penaltySeconds?: number;
    readonly penaltyReason?: number;
    readonly vacBanned?: number;
    readonly ranking?: PlayerRankingInfo;
    readonly commendation?: PlayerCommendationInfo;
    readonly medals?: PlayerMedalsInfo;
    readonly myCurrentEvent?: TournamentEvent;
    readonly myCurrentEventTeams?: TournamentTeam[];
    readonly myCurrentTeam?: TournamentTeam;
    readonly myCurrentEventStages?: TournamentEvent[];
    readonly surveyVote?: number;
    readonly activity?: AccountActivity;
    readonly playerLevel?: number;
    readonly playerCurXp?: number;
    readonly playerXpBonusFlags?: number;
    readonly rankings?: PlayerRankingInfo[];
    readonly owcaseid?: bigint;
}
// ...
export interface CMsgGCCStrike15v2PartySearch {
    readonly ver?: number;
    readonly apr?: number;
    readonly ark?: number;
    readonly grps?: number[];
    readonly launcher?: number;
    readonly gameType?: number;
}

export interface CMsgGCCStrike15v2PartySearchResults {
    readonly entries?: CMsgGCCStrike15v2PartySearchResults_Entry[];
}

export interface CMsgGCCStrike15v2PlayersProfile {
    readonly requestId?: number;
    readonly accountProfiles?: CMsgGCCStrike15v2MatchmakingGC2ClientHello[];
}

export interface CMsgGCCStrike15v2PremierSeasonSummary {
    readonly accountId?: number;
    readonly seasonId?: number;
    readonly dataPerWeek?: CMsgGCCStrike15v2PremierSeasonSummary_DataPerWeek[];
    readonly dataPerMap?: CMsgGCCStrike15v2PremierSeasonSummary_DataPerMap[];
}
// ...
export interface CMsgRecurringMissionSchema {
    readonly missions?: CMsgRecurringMissionSchema_MissionTemplateList[];
}

export interface CMsgRequestInventoryRefresh {
}

export interface CMsgRequestRecurringMissionSchedule {
}
// ...
export interface Cs2ClientHello {
    readonly version?: number;
    readonly socacheHaveVersions?: CMsgSOCacheHaveVersion[];
    readonly clientSessionNeed?: number;
    readonly clientLauncher?: number;
    readonly steamLauncher?: number;
}

export interface Cs2ClientWelcome {
    readonly version?: number;
    readonly gameData?: Uint8Array;
    readonly outofdateSubscribedCaches?: CMsgSOCacheSubscribed[];
    readonly uptodateSubscribedCaches?: CMsgSOCacheSubscriptionCheck[];
    readonly gameData2?: Uint8Array;
    readonly rtime32GcWelcomeTimestamp?: number;
    readonly currency?: number;
    readonly balance?: number;
    readonly balanceUrl?: string;
    readonly txnCountryCode?: string;
}
// ...
export interface Cs2TournamentPredictions {
    readonly eventid?: number;
}
// ...
export const Msg: {
    SOCacheSubscribed: MessageId;
    SOMultipleObjects: MessageId;
    SOCacheSubscriptionRefresh: MessageId;
    GCSetItemPosition: MessageId;
    GCVerifyCacheSubscription: MessageId;
    GCSetItemPositions: MessageId;
    GCApplySticker: MessageId;
    GCItemAcknowledged: MessageId;
    GCItemCustomizationNotification: MessageId;
    GCStoreGetUserData: MessageId;
    GCStoreGetUserDataResponse: MessageId;
    GCAdjustEquipSlotsManual: MessageId;
    GCAdjustEquipSlotsShuffle: MessageId;
    GCOpenCrate: MessageId;
    GCClientWelcome: MessageId;
    GCGameServerWelcome: MessageId;
    GCClientHello: MessageId;
    GCGameServerHello: MessageId;
    GCClientConnectionStatus: MessageId;
    MatchmakingStart: MessageId;
    MatchmakingStop: MessageId;
    MatchmakingClient2ServerPing: MessageId;
    MatchmakingGC2ClientUpdate: MessageId;
    MatchmakingGC2ServerReserve: MessageId;
    MatchmakingServerReservationResponse: MessageId;
    MatchmakingGC2ClientReserve: MessageId;
    MatchmakingClient2GCHello: MessageId;
    MatchmakingGC2ClientHello: MessageId;
    ClientRequestPlayersProfile: MessageId;
    PlayersProfile: MessageId;
    MatchEndRunRewardDrops: MessageId;
    MatchList: MessageId;
    MatchListRequestRecentUserGames: MessageId;
    MatchListRequestTournamentGames: MessageId;
    Server2GCClientValidate: MessageId;
    MatchListRequestTournamentPredictions: MessageId;
    AcknowledgePenalty: MessageId;
    ClientPlayerDecalSign: MessageId;
    PartySearch: MessageId;
    AccountRequestCoPlays: MessageId;
    ClientGCRankUpdate: MessageId;
    GetEventFavoritesRequest: MessageId;
    GetEventFavoritesResponse: MessageId;
    PremierSeasonSummary: MessageId;
    RequestRecurringMissionSchedule: MessageId;
    RecurringMissionSchema: MessageId;
    VolatileShopSubscribe: MessageId;
} = {
    SOCacheSubscribed: 24,
    SOMultipleObjects: 26,
    SOCacheSubscriptionRefresh: 28,
    GCSetItemPosition: 1001,
    GCVerifyCacheSubscription: 1005,
    GCSetItemPositions: 1077,
    GCApplySticker: 1086,
    GCItemAcknowledged: 1087,
    GCItemCustomizationNotification: 1090,
    GCStoreGetUserData: 2500,
    GCStoreGetUserDataResponse: 2501,
    GCAdjustEquipSlotsManual: 2531,
    GCAdjustEquipSlotsShuffle: 2532,
    GCOpenCrate: 2534,
    GCClientWelcome: 4004,
    GCGameServerWelcome: 4005,
    GCClientHello: 4006,
    GCGameServerHello: 4007,
    GCClientConnectionStatus: 4009,
    MatchmakingStart: 9101,
    MatchmakingStop: 9102,
    MatchmakingClient2ServerPing: 9103,
    MatchmakingGC2ClientUpdate: 9104,
    MatchmakingGC2ServerReserve: 9105,
    MatchmakingServerReservationResponse: 9106,
    MatchmakingGC2ClientReserve: 9107,
    MatchmakingClient2GCHello: 9109,
    MatchmakingGC2ClientHello: 9110,
    ClientRequestPlayersProfile: 9127,
    PlayersProfile: 9128,
    MatchEndRunRewardDrops: 9136,
    MatchList: 9139,
    MatchListRequestRecentUserGames: 9141,
    MatchListRequestTournamentGames: 9146,
    Server2GCClientValidate: 9153,
    MatchListRequestTournamentPredictions: 9160,
    AcknowledgePenalty: 9171,
    ClientPlayerDecalSign: 9185,
    PartySearch: 9191,
    AccountRequestCoPlays: 9193,
    ClientGCRankUpdate: 9194,
    GetEventFavoritesRequest: 9201,
    GetEventFavoritesResponse: 9203,
    PremierSeasonSummary: 9224,
    RequestRecurringMissionSchedule: 9225,
    RecurringMissionSchema: 9226,
    VolatileShopSubscribe: 9228,
};
// ...
export const Routes = {
    AccountCoPlays: {
        requestId: Msg.AccountRequestCoPlays,
        request: Proto.CMsgGCCStrike15v2AccountRequestCoPlays,
        responseId: Msg.AccountRequestCoPlays,
        response: Proto.CMsgGCCStrike15v2AccountRequestCoPlays
    } as GcRoute<CMsgGCCStrike15v2AccountRequestCoPlays, CMsgGCCStrike15v2AccountRequestCoPlays>,
    ClientHello: {
        requestId: Msg.GCClientHello,
        request: Proto.Cs2ClientHello,
        responseId: Msg.GCClientWelcome,
        response: Proto.Cs2ClientWelcome
    } as GcRoute<Cs2ClientHello, Cs2ClientWelcome>,
    EventFavorites: {
        requestId: Msg.GetEventFavoritesRequest,
        request: Proto.CMsgGCCStrike15v2GetEventFavoritesRequest,
        responseId: Msg.GetEventFavoritesResponse,
        response: Proto.CMsgGCCStrike15v2GetEventFavoritesResponse
    } as GcRoute<CMsgGCCStrike15v2GetEventFavoritesRequest, CMsgGCCStrike15v2GetEventFavoritesResponse>,
    MatchmakingHello: {
        requestId: Msg.MatchmakingClient2GCHello,
        request: Proto.CMsgGCCStrike15v2MatchmakingClient2GCHello,
        responseId: Msg.MatchmakingGC2ClientHello,
        response: Proto.CMsgGCCStrike15v2MatchmakingGC2ClientHello
    } as GcRoute<CMsgGCCStrike15v2MatchmakingClient2GCHello, CMsgGCCStrike15v2MatchmakingGC2ClientHello>,
    PartySearch: {
        requestId: Msg.PartySearch,
        request: Proto.CMsgGCCStrike15v2PartySearch,
        responseId: Msg.PartySearch,
        response: Proto.CMsgGCCStrike15v2PartySearchResults
    } as GcRoute<CMsgGCCStrike15v2PartySearch, CMsgGCCStrike15v2PartySearchResults>,
    PremierSeasonSummary: {
        requestId: Msg.PremierSeasonSummary,
        request: Proto.CMsgGCCStrike15v2PremierSeasonSummary,
        responseId: Msg.PremierSeasonSummary,
        response: Proto.CMsgGCCStrike15v2PremierSeasonSummary
    } as GcRoute<CMsgGCCStrike15v2PremierSeasonSummary, CMsgGCCStrike15v2PremierSeasonSummary>,
    RankUpdate: {
        requestId: Msg.ClientGCRankUpdate,
        request: Proto.CMsgGCCStrike15v2ClientGCRankUpdate,
        responseId: Msg.ClientGCRankUpdate,
        response: Proto.CMsgGCCStrike15v2ClientGCRankUpdate
    } as GcRoute<CMsgGCCStrike15v2ClientGCRankUpdate, CMsgGCCStrike15v2ClientGCRankUpdate>,
    RecentUserGames: {
        requestId: Msg.MatchListRequestRecentUserGames,
        request: Proto.CMsgGCCStrike15v2MatchListRequestRecentUserGames,
        responseId: Msg.MatchList,
        response: Proto.CMsgGCCStrike15v2MatchList
    } as GcRoute<CMsgGCCStrike15v2MatchListRequestRecentUserGames, CMsgGCCStrike15v2MatchList>,
    RecurringMissionSchedule: {
        requestId: Msg.RequestRecurringMissionSchedule,
        request: Proto.CMsgRequestRecurringMissionSchedule,
        responseId: Msg.RecurringMissionSchema,
        response: Proto.CMsgRecurringMissionSchema
    } as GcRoute<CMsgRequestRecurringMissionSchedule, CMsgRecurringMissionSchema>,
    RequestPlayersProfile: {
        requestId: Msg.ClientRequestPlayersProfile,
        request: Proto.CMsgGCCStrike15v2ClientRequestPlayersProfile,
        responseId: Msg.PlayersProfile,
        response: Proto.CMsgGCCStrike15v2PlayersProfile
    } as GcRoute<CMsgGCCStrike15v2ClientRequestPlayersProfile, CMsgGCCStrike15v2PlayersProfile>,
    StoreUserData: {
        requestId: Msg.GCStoreGetUserData,
        request: Proto.CMsgStoreGetUserData,
        responseId: Msg.GCStoreGetUserDataResponse,
        response: Proto.CMsgStoreGetUserDataResponse
    } as GcRoute<CMsgStoreGetUserData, CMsgStoreGetUserDataResponse>,
    TournamentGames: {
        requestId: Msg.MatchListRequestTournamentGames,
        request: Proto.CMsgGCCStrike15v2MatchListRequestTournamentGames,
        responseId: Msg.MatchList,
        response: Proto.CMsgGCCStrike15v2MatchList
    } as GcRoute<CMsgGCCStrike15v2MatchListRequestTournamentGames, CMsgGCCStrike15v2MatchList>,
    TournamentPredictions: {
        requestId: Msg.MatchListRequestTournamentPredictions,
        request: Proto.Cs2TournamentPredictions,
        responseId: Msg.MatchListRequestTournamentPredictions,
        response: Proto.Cs2TournamentPredictions
    } as GcRoute<Cs2TournamentPredictions, Cs2TournamentPredictions>,
    VolatileShop: {
        requestId: Msg.VolatileShopSubscribe,
        request: Proto.CMsgGCCStrike15v2VolatileShopSubscribe,
        responseId: Msg.VolatileShopSubscribe,
        response: Proto.CMsgGCCStrike15v2VolatileShopSubscribe
    } as GcRoute<CMsgGCCStrike15v2VolatileShopSubscribe, CMsgGCCStrike15v2VolatileShopSubscribe>,
} as const;